Modify Maven package command so it will complete.
Previous version would quietly fail and not build anything, resulting in
an image that can't run Zeppelin.

Changing to use -q with mvn doesn't make it entirely silent, but the
overall run of `make -C zeppelin install` produces 2387 lines of output
on my machine, which should fit in the Travis CI output.